refactor: character 라우터로 변경, 프론트 반영
