+++ packages/backend/src/character/dtos/base-character.dto.ts
