+++ packages/frontend/src/auth/hooks/useCharacter.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import {
-  AuthControllerCharacterSelectParams,
+  CharacterControllerCharacterSelectParams,
   GetCharacter,
 } from '@rimgosu/libs';
 import { ApiSingleton } from '../../common/apiSingleton';
@@ -11,7 +11,7 @@ export const useCharacter = () => {
     ApiResponse<GetCharacter[]>
   > => {
     return await ApiSingleton.getInstance()
-      .auth.authControllerGetCharacters()
+      .characters.characterControllerGetCharacters()
       .then((res) => ({ data: res.data }))
       .catch(async (error: Response) => {
         return { error: (await error.json()) as ApiErrorType };
@@ -19,10 +19,10 @@ export const useCharacter = () => {
   }, []);
 
   const selectCharacter = async (
-    params: AuthControllerCharacterSelectParams,
+    params: CharacterControllerCharacterSelectParams,
   ): Promise<ApiResponse<void>> => {
     return await ApiSingleton.getInstance()
-      .auth.authControllerCharacterSelect(params)
+      .characters.characterControllerCharacterSelect(params)
       .then((res) => ({ data: res.data }))
       .catch(async (error: Response) => {
         return { error: (await error.json()) as ApiErrorType };
